requirements.txt
agora_token_builder==1.0.0
asgiref==3.8.1
boto3==1.35.6
botocore==1.35.6
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
distlib==0.3.8
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
gunicorn==23.0.0
idna==3.7
jmespath==1.0.1
ndg-httpsclient==0.5.1
packaging==24.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pusher==3.3.2
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
